Basic project setup: allocator lib with demo.
Here, we have a basic setup of one of the
libraries that we will be creating in this
project.

We want to write an allocator, several different
containers like vectors, stacks, linked_lists,
etc. We'll have each of the different libraries in
their own folders respectively.

Each of the libraries will be independent. There
will be the src and include of the library in the
root directory of that project. Also in the root
directory will be the demos and tests directories.
The demos directory will then contain one or more
demo projects, each of which have thier own CMake
files, along with their own src and includes.

It should be possible to build just the libraries,
just a specific set of libraries, without having
to build all the tests and examples along with
them. There is still work to do to achieve this
(since an add_subdirectory command in the root
CMakeLists.txt is not likely to be the correct
solution).