Extend the logger_print_helper's capabilites
Now, the print_helper fn can also specify the
value denoting the nature of the log message -
whether it is an error, trace, debug stmt, or
an info statement.

Design notes:
* We didn't want to pass a string, since then
that is a little bit arbitrary - the kinds of
log messages should be limited to several
different key values. Hence, the use of enums.
(Also, this means that the logger doesn't have
to do any string length checking and such on
values passed in).

* We needed to do this because otherwise, if
we decided to create a logger_info or a
logger_debug function, then they would all
have the same code that our logger_error had,
since the logger_print fn originally couldn't
dynamically change that it was printing
"[TRACE]".