Create the heapsort module, incorporate logger.
The logger here requires that there be an
initialization function in order to correctly
operate. (Unfortunately, the stdout value is not
a constant, and so must be assigned at runtime,
and not at compile time. Hence, the module
initialize function.

To test out the logger with at least something,
the is_sorted fn was also created.