Implement the is_heap fn in the demo.
This function introduces a lot of output, and is
a prime example for why we need some filtering
functionality in our logger. Since the is_heap
fn makes frequent use of the parent fn, we get
a lot of parent calls, and hence a lot of
printouts by parent. It would be nice if we could
say that we only ever want to hear from parent
when there is an error or something, but no more
traces.