Fix backwards pathing in stack's CMakeLists.txt
Since we don't want backwards pathing in our
stack's CMakeLists.txt, and we need vector's
include, we need a CMakeLists in the containers
directory that creates a variable that has the
path of the container's directory in it.

Since the project root has changed, we need to
place the enable_testing statement in the new
CMakeLists.txt so that we don't have to go down
to the vector directory to build tests.

Note: There is more configuring to do, since we
want to be able to selectively build the tests
and components in general.