Create the init test for the stack type.
Note: How do we test this right, since to check
whether things happen right, we always need to be
checking against the underlying data structure.

It seems that the only thing we can really do is
run through all the possible fns that operate on
a stack and see if the results that we expect.