Add specs for vector and stack pop fns.
It's important to know what happens in the
case that someone does a pop on an empty
stack or a pop_back on an empty vector.

Since in the cppreference it says that they're
undefined, when the container is empty, we
decide to define them for this implementation
as having no effect.