Bugfix: list return values and struct defns
In our list implementation, we accidentally
pollute the namespace by declaring the node
structure, when that header is set up to
use void pointers to avoid exposing the
list_node internal structure. (Another possible
implementation is to declare a list_impl
and have a pointer to it), and list_impl
is defined in the the list.c file).

The second bugfix was that the fn that
creates list nodes didn't actually return
the pointer to the node, but amazingly,
the way C works, the correct value would
be left in the register or something, which
would allow the caller to get the pointer
they expected. This was fixed.